feat: add weakness command, type effectiveness for pokemon details
