+++ package.json
@@ -33,6 +33,13 @@
       "subtitle": "Pokédex",
       "description": "A move, also known as an attack or technique, is the skill Pokémon primarily use in battle.",
       "mode": "view"
+    },
+    {
+      "name": "weakness",
+      "title": "Weaknesses",
+      "subtitle": "Pokédex",
+      "description": "Discover the types of moves that are particularly effective against a specific Pokémon.",
+      "mode": "view"
     }
   ],
   "preferences": [
@@ -84,8 +91,8 @@
     }
   ],
   "dependencies": {
-    "@raycast/api": "^1.68.1",
-    "axios": "^1.6.7",
+    "@raycast/api": "^1.82.5",
+    "axios": "^1.7.7",
     "json2md": "^2.0.1",
     "lodash.debounce": "^4.0.8",
     "lodash.groupby": "^4.6.0"
@@ -94,15 +101,15 @@
     "@types/json2md": "^1.5.4",
     "@types/lodash.debounce": "^4.0.9",
     "@types/lodash.groupby": "^4.6.9",
-    "@types/node": "~20.11.20",
-    "@types/react": "^18.2.59",
+    "@types/node": "~22.5.5",
+    "@types/react": "^18.3.5",
     "@typescript-eslint/eslint-plugin": "^7.1.0",
     "@typescript-eslint/parser": "^7.1.0",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
-    "prettier": "^3.2.5",
-    "react-devtools": "^4.28.4",
-    "typescript": "^5.3.3"
+    "prettier": "^3.3.3",
+    "react-devtools": "^5.2.0",
+    "typescript": "^5.6.2"
   },
   "scripts": {
     "build": "ray build -e dist",
